requirements.txt
fastapi
uvicorn
langchain
langchain-ollama
langchain-postgres
langchain-community
psycopg[binary]
redis
pypdf
python-dotenv